separate data fetching logic into RibbonDataProvider component * RibbonDataProvider takes a callback function as children component * Ribbon can be used on its own or with RibbonDataProvider
