+++ src/index.js
@@ -1,3 +1,8 @@
 import Ribbon from './Ribbon';
+import RibbonDataProvider from './RibbonDataProvider';
 
 export default Ribbon;
+
+export {
+  RibbonDataProvider
+};
